refactor fuse impl of plane_topk_merge and plane_argtopk_merge
